Organize tests into feature namespaces
Move and rename multiple test files into feature-specific folders and update their namespaces to match the new paths. Affected areas include Archive, Commands, Host (from Helpers), Messaging, Resources (from Tools/Documents), Server, Settings, Spreadsheet (from Tools), UserInterface, WebHost (from WebView), and WebView. This reorganizes test layout for clearer project structure and does not change test logic.
