+++ Source/Tests/Commands/CommandTests.cs
@@ -7,7 +7,7 @@ using Celbridge.UserInterface.Services;
 using Celbridge.Workspace;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Celbridge.Tests;
+namespace Celbridge.Tests.Commands;
 
 public class TestCommand : CommandBase
 {
