+++ Source/Tests/Resources/PathValidatorTests.cs
@@ -1,6 +1,6 @@
 using Celbridge.Resources.Services;
 
-namespace Celbridge.Tests;
+namespace Celbridge.Tests.Resources;
 
 [TestFixture]
 public class PathValidatorTests
